+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # dexter 1.6.2
 
+* support for .data$ and .env$ pronouns in predicates
+
 * Bugfix when person_id or booklet_id are numeric, the outputted values were altered in ability, plausible values and plausible scores. 
 
 # dexter 1.6.1